FM2-400: Moved LastnResults.java to utils folder
